Update power-meter.yml
Fixed a bug that left the some sensors unknown (nan) after a reboot. Only resetting the totals seems to work. This happened because I didn't check whether the pulse_meter state might happen to be nan - in the process_power script - resulting in adding a nan to the counters... and as we all know: any number + nan = nan.

Turns out the pulse meter manages to trigger two total: on_value events before the pulse meter state is published, after reboot.

The fix is ignoring all the triggers where the power state is nan, which should be the two first triggers - which again will delay the counting one pulse.

This is a temporary fix as I realize the script has to be rewritten.